inscrie-te.html - full temp
This page now shows that you can't sign up anymore. This is the second one (first was just an error), and the form is currently hidden with display:none. Next step is to delete the entire form code (i do this for 10 more people that we have to sign up).